requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.0
gunicorn==20.1.0
kombu==5.4.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13